Feature/diet generation

+++ src/main/java/com/nutriplus/NutriPlusBack/domain/food/Food.java
@@ -24,7 +24,7 @@ public class Food extends FoodModel {
     }
     //New Food
     public Food(String foodNameValue, String foodGroupValue, double measureTotalGramsValue, String measureTypeValue,
-         int measureAmountValue, NutritionFacts nutritionFactsValue){
+         Double measureAmountValue, NutritionFacts nutritionFactsValue){
         super();
         foodName            = foodNameValue;
         foodGroup           = foodGroupValue;
@@ -37,7 +37,7 @@ public class Food extends FoodModel {
     }
     //New created food (requires nutritionist parameter)
     public Food(UserCredentials owner, String foodNameValue, String foodGroupValue, double measureTotalGramsValue, String measureTypeValue,
-                int measureAmountValue, NutritionFacts nutritionFactsValue){
+                Double measureAmountValue, NutritionFacts nutritionFactsValue){
         super();
         foodName            = foodNameValue;
         foodGroup           = foodGroupValue;
@@ -80,7 +80,7 @@ public class Food extends FoodModel {
     public void setFoodGroup(String foodGroupValue)                   { foodGroup = foodGroupValue; }
     public void setMeasureTotalGrams(double measureTotalGramsValue)   { measureTotalGrams = measureTotalGramsValue; }
     public void setMeasureType(String measureTypeValue)               { measureType = measureTypeValue; }
-    public void setMeasureAmount(int measureAmountValue)              { measureAmount = measureAmountValue; }
+    public void setMeasureAmount(Double measureAmountValue)              { measureAmount = measureAmountValue; }
     public void setNutritionFacts(NutritionFacts nutritionFactsValue) {
         if(nutritionFacts == null){
             nutritionFacts = new NutritionFacts();
@@ -133,6 +133,6 @@ public class Food extends FoodModel {
     public String getFoodGroup()              { return foodGroup; }
     public double getMeasureTotalGrams()      { return measureTotalGrams; }
     public String getMeasureType()            { return measureType; }
-    public int getMeasureAmount()             { return measureAmount; }
+    public Double getMeasureAmount()             { return measureAmount; }
     public NutritionFacts getNutritionFacts() { return nutritionFacts; }
 }
